+++ com.unity.render-pipelines.high-definition/Editor/Sky/HDRISky/HDRISkyEditor.cs
@@ -11,8 +11,6 @@ namespace UnityEditor.Rendering.HighDefinition
         : SkySettingsEditor
     {
         SerializedDataParameter m_hdriSky;
-        SerializedDataParameter m_DesiredLuxValue;
-        SerializedDataParameter m_IntensityMode;
         SerializedDataParameter m_UpperHemisphereLuxValue;
 
         RTHandle m_IntensityTexture;
@@ -23,13 +21,13 @@ namespace UnityEditor.Rendering.HighDefinition
         {
             base.OnEnable();
 
+            m_EnableLuxIntensityMode = true;
+
             // HDRI sky does not have control over sun display.
             m_CommonUIElementsMask = 0xFFFFFFFF & ~(uint)(SkySettingsUIElement.IncludeSunInBaking);
 
             var o = new PropertyFetcher<HDRISky>(serializedObject);
             m_hdriSky = Unpack(o.Find(x => x.hdriSky));
-            m_DesiredLuxValue = Unpack(o.Find(x => x.desiredLuxValue));
-            m_IntensityMode = Unpack(o.Find(x => x.skyIntensityMode));
             m_UpperHemisphereLuxValue = Unpack(o.Find(x => x.upperHemisphereLuxValue));
 
             m_IntensityTexture = RTHandles.Alloc(1, 1, colorFormat: GraphicsFormat.R32G32B32A32_SFloat);
@@ -73,37 +71,12 @@ namespace UnityEditor.Rendering.HighDefinition
             EditorGUI.BeginChangeCheck();
             {
                 PropertyField(m_hdriSky);
-
-                EditorGUILayout.Space();
-
-                PropertyField(m_IntensityMode);
+                base.CommonSkySettingsGUI();
             }
             if (EditorGUI.EndChangeCheck())
             {
                 GetUpperHemisphereLuxValue();
             }
-
-            if (m_IntensityMode.value.GetEnumValue<SkyIntensityMode>() == SkyIntensityMode.Lux)
-            {
-                EditorGUI.indentLevel++;
-                PropertyField(m_DesiredLuxValue);
-                // Hide exposure and multiplier
-                m_CommonUIElementsMask &= ~(uint)(SkySettingsUIElement.Exposure | SkySettingsUIElement.Multiplier);
-                m_CommonUIElementsMask |= (uint)SkySettingsUIElement.IndentExposureAndMultiplier;
-
-                // Show the multiplier
-                EditorGUILayout.HelpBox(System.String.Format("Upper hemisphere lux value: {0}\nAbsolute multiplier: {1}",
-                    m_UpperHemisphereLuxValue.value.floatValue,
-                    (m_DesiredLuxValue.value.floatValue / m_UpperHemisphereLuxValue.value.floatValue)
-                ), MessageType.Info);
-                EditorGUI.indentLevel--;
-            }
-            else
-            {
-                m_CommonUIElementsMask |= (uint)(SkySettingsUIElement.Exposure | SkySettingsUIElement.Multiplier);
-            }
-
-            base.CommonSkySettingsGUI();
         }
     }
 }