+++ com.unity.render-pipelines.high-definition/Runtime/Sky/HDRISky/HDRISkyRenderer.cs
@@ -23,13 +23,10 @@ namespace UnityEngine.Rendering.HighDefinition
         public override void RenderSky(BuiltinSkyParameters builtinParams, bool renderForCubemap, bool renderSunDisk)
         {
             var hdriSky = builtinParams.skySettings as HDRISky;
-            float luxMultiplier = hdriSky.desiredLuxValue.value / hdriSky.upperHemisphereLuxValue.value;
-            float multiplier = (hdriSky.skyIntensityMode.value == SkyIntensityMode.Exposure) ? hdriSky.multiplier.value : luxMultiplier;
-            float exposure = (hdriSky.skyIntensityMode.value == SkyIntensityMode.Exposure) ? GetExposure(hdriSky, builtinParams.debugSettings) : 1;
             float phi = Mathf.Deg2Rad * -hdriSky.rotation.value; // -rotation to match Legacy...
 
             m_SkyHDRIMaterial.SetTexture(HDShaderIDs._Cubemap, hdriSky.hdriSky.value);
-            m_SkyHDRIMaterial.SetVector(HDShaderIDs._SkyParam, new Vector4(exposure, multiplier, Mathf.Cos(phi), Mathf.Sin(phi)));
+            m_SkyHDRIMaterial.SetVector(HDShaderIDs._SkyParam, new Vector4(GetSkyIntensity(hdriSky, builtinParams.debugSettings), 0.0f, Mathf.Cos(phi), Mathf.Sin(phi)));
 
             using (new ProfilingSample(builtinParams.commandBuffer, "Draw sky"))
             {